requirements.txt
django
django-cors-headers
django-filter
djangorestframework
# gunicorn
numpy
pandas
Pillow
psycopg2-binary
python-decouple
